Extracting server actions from api routes
