+++ src/lib/server-actions/recipes/getLikedRecipes.ts
@@ -1,0 +1,9 @@
+import DBClient from "@/persistence/DBClient";
+
+const prisma = DBClient.getInstance().prisma;
+
+export const getLikedRecipes = async (userId: string) => {
+  return await prisma.userRecipeFavorite.findMany({
+    where: { userId: userId },
+  });
+};
